# swagger.json
{
    "swagger": "2.0",
    "info": {
    "version": "1.0.0",
    "title": "Fairer Coffee",
    "description": "Application API Guide.",
    "license": {
    "name": "Sam Obrien - IOT"
    }
    },
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/orders/{order_id}": {
            "get": {
                "tags": [
                    "Orders Functionality"
                ],
                "summary": "Retrieves an Order by ID",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "description": "ID of the Order"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This will return a 200 response and the corresponding Order."
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/orders/order_trace/{order_trace_value}": {
            "get": {
                "tags": [
                    "Orders Functionality"
                ],
                "summary": "Retrieves an Order by Order Trace Value",
                "parameters": [
                    {
                        "name": "order_trace_value",
                        "in": "path",
                        "required": true,
                        "description": "Order Trace Value",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "This will return a 200 response and the corresponding Order."
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/orders/import": {
            "post": {
                "tags": [
                    "Orders Functionality"
                ],
                "summary": "Import a New Order",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "query",
                        "description": "Order ID",
                        "required": true
                    },
                    {
                        "name": "order_date",
                        "in": "query",
                        "description": "Order Date",
                        "required": true
                    },
                    {
                        "name": "order_customer",
                        "in": "query",
                        "description": "Order Customer",
                        "required": true
                    },
                    {
                        "name": "order_total",
                        "in": "query",
                        "description": "Order Total",
                        "required": true
                    },
                    {
                        "name": "order_count",
                        "in": "query",
                        "description": "Order Count",
                        "required": true
                    },
                    {
                        "name": "order_volume",
                        "in": "query",
                        "description": "Order Volume",
                        "required": true
                    },
                    {
                        "name": "order_trace",
                        "in": "query",
                        "description": "Order Trace",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order imported successfully"
                    },
                    "400": {
                        "description": "Bad request. Invalid data provided."
                    }
                }
            }
        }
    }
}